+++ src/Local_Development/Init.php
@@ -29,7 +29,7 @@ class Init {
 	public function __construct() {
 		$config = get_site_option( 'local_development', [] );
 		$config = $this->get_vcs_checkouts( $config );
-		$config = $this->get_github_updater_running( $config );
+		$config = $this->modify_options( $config );
 		add_action(
 			'init',
 			function () use ( $config ) {
@@ -77,15 +77,18 @@ class Init {
 	}
 
 	/**
-	 * Modify option if GitHub Updater is running.
+	 * Modify options.
 	 *
 	 * @param array $config Plugin options.
 	 *
 	 * @return array
 	 */
-	private function get_github_updater_running( $config ) {
+	private function modify_options( $config ) {
 		if ( is_plugin_active( 'github-updater/github-updater.php' ) ) {
 			$config['extras']['enable_git_icons'] = '-1';
+		}
+		if ( defined( 'WP_DISABLE_FATAL_ERROR_HANDLER' ) && WP_DISABLE_FATAL_ERROR_HANDLER ) {
+			$config['extras']['bypass_fatal_error_handler'] = '-1';
 		}
 		return $config;
 	}