+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* set disabled checked option if `WP_DISABLE_FATAL_ERROR_HANDLER` is true
 
 #### 2.5.6 / 2020-07-10
 * fix logic for no setting